Move away from :certificate face, use Puppet::SSL
There is some crazy behavior when using the certificate face. Race
condition observable sometimes when you fire off a `puppet apply` run
and the certificate by the name of the $certname setting doesn't exist
yet. Eric Sorenson suggested that bypassing the certificate face might
make things better. No proof of that but here's an implementation to
give the hypothesis a whirl.