Ignore `certificate unknown` error when cleaning cert
This can happen if the CA doesn't have a copy of your signed cert.
Perhaps it was already cleaned, or perhaps it was generated on a standby
CA server by mistake, (and then overwritten).

Unfortunately, the exception isn't more specific than
`OpenSSL::SSL::SSLError` and doesn't contain a code, so checking if the
message includes `certificate unknown` is the best I can do.

Fixes #13